Math: redo Matrix4::normalMatrix() as an inverse transpose.
The article this API was originally based on assumes a scenario which
just *isn't* matching usual practices here, giving wrong results. Too
bad I didn't spend more time questioning the proof there and just
blindly assumed it's correct because everyone said so.

Won't be typing all that reasoning again in the commit message, see the
changelog entry and the comment in the test for details.