TextureTools: add a DistanceField overload taking a framebuffer.
So one can directly read it back on GLES without having to wrap the
texture in a framebuffer again.

This change also puts the framebuffer completeness check *before* the
clear() and bind() which makes it no longer emit a GL error. The error
is still silent though, which isn't nice. Gotta fix that eventually as
well.