Trade: add *{Importer,Converter}Flag::Quiet.
A counterpart / inverse to Verbose, is meant to be used by plugins to
suppress all warnings.

This flag was already used by the ShaderConverter APIs. Fonts don't have
any flags (and don't have any verbose output or warnings at the moment
either), so there it isn't added; audio importers are in a maintenance
mode with no new features added as I'll be eventually merging them with
the general importer interface in Trade.