CMake: allow dynamic builds on Android and Emscripten.
Same as in Corrade, they still default to static because that's the
least painful.

Plugins however still need to be forced to be built static on platforms
where the PluginManager doesn't have any support for dynamic plugins
implemented. The logic is now only simplified, to not have to change it
again when e.g. Android gets the needed bits implemented in
PluginManager.

Co-authored-by: Vladimír Vondruš <mosra@centrum.cz>