[wip] Platform: create a Sdl2ApplicationWindow base class.
Mostly just code motion and related documentation updates because
Doxygen is such a crap that it can link to functions defined in a base
class but not types! FFS.

No actual multi-window support yet, that'll be in the next commits to
avoid this noise obscuring them.

TODO: Sdl2Window instead?
TODO: unsure about certain APIs, whether they should be per-window or
not
 - contextless / gl / vulkan?
 - mouse warp?
 - ...?