MeshTools: deprecate owned() in favor of copy().
A bit sad it took me three years to invent the right name for this
utility, heh. Also moving it together with others to a new
MeshTools/Copy.h header because *this* is the mainly useful API, not
reference() / mutableReference().

MaterialTools and SceneTools will get similar copy() APIs doing the same
thing.