Trade: don't unwrap custom attribute/field names when passing to do*().
Originally I thought I'd save plugins some time if I just give them the
custom indices directly, without being wrapped in a SceneField /
MeshAttribute enum. But in practice that didn't really save anything,
made the interfaces more error-prone due to there being no concrete type
anymore, and all code that delegates to nested importers or converters
had to re-wrap the IDs again, which is *again* error-prone.

Bumping the interface strings because this is a breaking change for the
implementations. Not for users tho, there nothing changes.