Trade: SceneFieldData::size() should be a std::size_t, not UnsignedLong.
Not sure what was I thinking here -- if a field size wouldn't fit into a
32bit number, it won't fit into memory of a 32bit system anyway, so
there's no real use for the size to be returned as 64bit always.
Internally it *is* stored as a 64bit number, yes, to have compatible
binary layout on 32bit and 64bit systems, but that doesn't mean the
public API should return that too. And SceneData::fieldSize() is
std::size_t, so this feels really like an accidental brainfart.

The changed return type also means a lot of existing code doesn't need
to do any explicit casting to std::size_t anymore. Yay.