debuggers/magnum.natvis: display the angle in Deg/Rad nicely (along w/ equivalent Rad/Deg in parentheses)
There are two variants per angle type, so a total of four entries.
There are separate entries for Float & Double variants because in order to calculate the equivalent value for the given angle type in the other type,
I used the same exact values for the multipliers as the ones in Magnum::Math (utilizing Math::Constants<T>::pi()).

Also I used the format specifier 'h' (hexadecimal integer) for Double variants, as it shows a lot more digits, which may be more desirable when working with double data types for angles.