Math: remove std::tuple usage from Algorithms::svd().
There's Containers::Triple now for this instead. Printing a message to
Error in case of a failure might have made sense back in 2010, but now
it absolutely doesn't, so it's additionally wrapped in an Optional now.

Also looks like the actual use without "convenient" std::tie() is a lot
less verbose. Haha.