MagnumFont{,Converter}: actual proper tests.
Need some regression tests for the upcoming glyph cache rework as
otherwise it'll be too miserable. It now fills the glyph cache image
with some non-trivial data and compares to them, and checks the filled
glyphs for the created glyph cache too.

This also fixes a regression introduced when porting away from STL in
47a1295ab80d33210abae03fe03ed6953fa949b2, where UTF-8 layouting was
reporting extra glyphs at the end. Now UTF-8 support is properly tested
both in the MagnumFontConverter plugin and MagnumFont.