Text: make GlyphCache movable.
And thus the DistanceFieldGlyphCache subclass as well. The
deinlined destructor wasn't really needed as the GL::Texture has its
destructor deinlined as well, so it wouldn't cause too much extra work
for the compiler to have it implicit.

Also, I suspect the destructor was just a leftover from when there was
no AbstractGlyphCache base.