Shaders: update DistanceFieldVector test files with fixed SDFs.
The output is now *much* closer to the Vector shader output (mean
difference went from ~6 to 1.79), and in the multi-draw case it's
clearly visible that it's no longer slightly weirdly uneven.

The test files are now also much smaller as the originals were created
before RLE was implemented in TgaImageConverter.